requirements.txt
alembic==1.13.1
annotated-types==0.6.0
async-timeout==4.0.3
Babel==2.15.0
bcrypt==4.1.2
blinker==1.7.0
boto3==1.34.11
botocore==1.34.11
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
elastic-transport==8.11.0
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
email_validator==2.1.1
environs==10.0.0
Flask==3.0.3
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fpdf==1.7.2
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.6
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
Mako==1.3.3
MarkupSafe==2.1.5
marshmallow==3.20.1
opensearch-dsl==2.1.0
opensearch-py==2.4.2
packaging==23.2
passlib==1.7.4
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
requests-aws4auth==1.2.3
rsa==4.9
s3transfer==0.10.0
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.25
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.2
WTForms==3.1.2
